Added docs for X11 key remapping (#33)
* Added docs for X11 key remapping

Lite-XL does not seem to be aware of any keyboard re-mapping done via the usual tools (such as `xkbset` or the GNOME setting).

I assume this is because Lite is built on SDL, which is getting its keyboard input via some mechanism that is bypassing the usual X11 drudgery?

Anywhoo, as a lifelong Caps -> Ctrl remapping this is kind of a big deal. Fortunately there's a pretty straightforward workaround. This is a big enough deal that I think it'll cause a lot of people to bounce off Lite-XL pretty quickly.

I'm not sure this is the best place in the docs but I think it's worth documenting *somewhere*. Happy to move locations or wording as suggested.

* Various PR feedback

* Moved everything inside a "tip" admonition
* Used the proper name "Lite XL"
* Suggest remapping via the Lite XL config

* Fix typo

* Rephrase a little bit

---------

Co-authored-by: takase1121 <20792268+takase1121@users.noreply.github.com>